requirements.txt
flask
magic
pdfplumber
datetime
flask_cors
os
functools
csv
werkzeug.exceptions
re
werkzeug.utils